Move DisplayDiffHunk into hunk_diff module
